+++ HISTORY.rst
@@ -20,6 +20,7 @@ Unreleased
 * ALCN-980: Translate journal "Wage" and "Miscellaneous Operations" and remove duplicate journal
 
 **Bugfixes**
+* There is no fridge product category. This is a route selectable on the product.
 
 * Allow to create a lot at reception with standard interface
 * Fix new flake8 errors